requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.5
drf-spectacular>=0.26
mysqlclient>=2.1
python-dotenv>=1.0
